Refactor gqls
