+++ src/graphql/gqls/apiKey.ts
@@ -1,0 +1,12 @@
+import { gql } from 'graphql-request';
+
+export const API_KEYS = gql`
+  query SchemaSyncApiKeys($first: Int!, $offset: Int!, $titles: [String!]) {
+    apiKeys(first: $first, offset: $offset, filterBy: { titles: $titles }) {
+      nodes {
+        id
+        title
+      }
+    }
+  }
+`;
