+++ src/graphql/gqls/dataPoint.ts
@@ -1,0 +1,96 @@
+import { gql } from 'graphql-request';
+
+export const DATA_POINTS = gql`
+  query SchemaSyncDataPoints($dataSiloIds: [ID!], $first: Int!, $offset: Int!) {
+    dataPoints(
+      filterBy: { dataSilos: $dataSiloIds }
+      first: $first
+      offset: $offset
+    ) {
+      totalCount
+      nodes {
+        id
+        title {
+          defaultMessage
+        }
+        description {
+          defaultMessage
+        }
+        name
+        purpose
+        category
+        actionSettings {
+          type
+          active
+        }
+        dbIntegrationQueries {
+          query
+          suggestedQuery
+          requestType
+        }
+      }
+    }
+  }
+`;
+
+export const SUB_DATA_POINTS = gql`
+  query SchemaSyncDataPoints(
+    $dataPointIds: [ID!]
+    $first: Int!
+    $offset: Int!
+  ) {
+    subDataPoints(
+      filterBy: { dataPoints: $dataPointIds }
+      first: $first
+      offset: $offset
+    ) {
+      totalCount
+      nodes {
+        id
+        name
+        description
+        purposes {
+          name
+          purpose
+        }
+        categories {
+          name
+          category
+        }
+      }
+    }
+  }
+`;
+
+export const UPDATE_OR_CREATE_DATA_POINT = gql`
+  mutation SchemaSyncUpdateOrCreateDataPoint(
+    $dataSiloId: ID!
+    $name: String!
+    $title: String
+    $description: String
+    $category: DataCategoryType
+    $purpose: ProcessingPurpose
+    $querySuggestions: [DbIntegrationQuerySuggestionInput!]
+    $enabledActions: [RequestActionObjectResolver!]
+    $subDataPoints: [DataPointSubDataPointInput!]
+  ) {
+    updateOrCreateDataPoint(
+      input: {
+        dataSiloId: $dataSiloId
+        name: $name
+        title: $title
+        description: $description
+        querySuggestions: $querySuggestions
+        category: $category
+        purpose: $purpose
+        enabledActions: $enabledActions
+        subDataPoints: $subDataPoints
+      }
+    ) {
+      dataPoint {
+        id
+        name
+      }
+    }
+  }
+`;
